add test timeout

+++ src/week3/test/normal/fibonacciWorker.js
@@ -1,6 +1,16 @@
 const { Worker } = require("worker_threads");
 
-const runFibonacciWorker = (num) => {
+/**
+ *
+ * @typedef {Object} Options
+ * @property {number} timeout
+ *
+ * @param {number} num
+ * @param {Options} options
+ *
+ * @returns {Promise<number>}
+ */
+const runFibonacciWorker = (num, options) => {
   return new Promise((resolve, reject) => {
     const worker = new Worker("./src/week3/test/normal/fibonacci.js", {
       workerData: { num },
@@ -11,10 +21,12 @@ const runFibonacciWorker = (num) => {
       if (code !== 0) reject(new Error(`stopped with  ${code} exit code`));
     });
 
-    setTimeout(() => {
-      resolve("Test timeout");
-      worker.terminate();
-    }, 1000);
+    if (options?.timeout) {
+      setTimeout(() => {
+        resolve("Test timeout");
+        worker.terminate();
+      }, options.timeout);
+    }
   });
 };
